requirements.txt
# openai~=1.78.1
langchain
langchain-core
langchain-openai
python-dotenv